% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_na_data.R
\name{check_na_data}
\alias{check_na_data}
\title{Check NAs in species data}
\usage{
check_na_data(
  species,
  species_column_name = "spec",
  species_gpkg = "data/species.gpkg",
  env_layers,
  remove = FALSE
)
}
\arguments{
\item{species}{short species name (character) used in project. For example:
satmon for Satureja montana, aursax for Aurinia saxatilis etc}

\item{species_column_name}{column name (character) as in .gpkg file from
where species data will be extracted. Default value is "spec"}

\item{env_layers}{environmental layers (raster stack) that is the result
of running create_env_stack function}

\item{remove}{optional value (boolean) to choose if removal of NAs will
also be performed}
}
\value{
Data table object
}
\description{
Check for NAs in species data by extracting data from environmental raster
layers. If argument remove is TRUE the function can automatically remove
NAs from the species dataset.
}
