% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_env_stack.R
\name{create_env_stack}
\alias{create_env_stack}
\title{Create raster stack from environmental data}
\usage{
create_env_stack(folder_in, env_subset = c())
}
\arguments{
\item{folder_in}{filename (character) or Raster object. Supported file types
are the "native" raster package format and those that can be read via rgdal}

\item{env_subset}{vector of filenames without extension (character) that
will be processed and stacked}
}
\value{
Raster stack
}
\description{
Create raster stack from environmental data so that raster stack by
providing folder where raster are stored and optionally subset of those
raster that needs processing. It is just a wrapper around raster::stack()
and used when you have subset of variables to try so that you can quickly
check for NAs with `check_na_data()` function or value extraction when
generating absences and psudoabsences
}
