% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_models.R
\name{create_models}
\alias{create_models}
\title{Create species distribution models}
\usage{
create_models(
  data,
  method,
  eval_metric_quality_threshold = NULL,
  nb_run_eval,
  path_to_bin = "bin/maxent.jar"
)
}
\arguments{
\item{data}{(object), Biomod formatted data  that is produced as result from
BIOMOD_FormattingData() function}

\item{method}{(character), modelling  method used to create model}

\item{eval_metric_quality_threshold}{(number) If not NULL, the minimum scores below
which models will be excluded of the ensemble-models building}

\item{nb_run_eval}{(number), number of Evaluation runs}

\item{path_to_bin}{(character), the link to maxent.jar file (the bin folder in
working directory by default)}
}
\description{
This function runs biomod2 and create species distributions models based on
user inputs. As prerequisite BIOMOD_FormatingData() needs to be run as well
as environmental layers stack. This function just wraps around biomod2
functions BIOMOD_Modeling() and BIOMOD_EnsembleModeling() to produce models
with consistent naming.
}
