% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_rasters.R
\name{prepare_rasters}
\alias{prepare_rasters}
\title{Prepare rasters}
\usage{
prepare_rasters(folder, extent)
}
\arguments{
\item{folder}{filename (character) or Raster object. Supported file types
are the ’native’ raster package format and those that can be read via rgdal}

\item{extent}{Extent object, or any object from which an Extent object can
be extracted}
}
\value{
RasterStack
}
\description{
Prepare environmental raster layers by loading them from the folder and
cropping to the project extent. In that process rasters are aligned to
the same grid and extent and converted to ASCII rasters.
}
