% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_species_data.R
\name{prepare_species_data}
\alias{prepare_species_data}
\title{Prepare species data}
\usage{
prepare_species_data(species, species_column_name = "spec", mask, env_layers)
}
\arguments{
\item{species}{(character), name used in project. For example:
satmon for Satureja montana, aursax for Aurinia saxatilis etc.}

\item{species_column_name}{(character), column name  as in .gpkg file from
where species data will be extracted. Default value is "spec"}

\item{mask}{sf polygon object with specific areas from where pseudoabsences
will be extracted}

\item{env_layers}{(raster stack), environmental layers  that is the result
of running create_env_stack function}
}
\value{
Data table object
}
\description{
Prepare species data by creating presence and/or pseudoabsence datasets for
modeling. In that process, first the NAs are removed, then presence points
are thinned to remove points that are too close together (less than 1.5km
when) modeling with ~1km pixel size. It is assumed that in wd the directory
containing Geopackage file with species data is present.
}
